Auto generate ids

+++ app/server.h
@@ -96,7 +96,7 @@ typedef struct {
     // streams
     s8 stream_key;
     i64 last_id_ms;
-    i32 last_id_seqn;
+    i64 last_id_seqn;
 } ServerContext;
 
 void* connection_handler(void* arg);
